# perf: add basic benchmark to demonstrate overhead